fix(KFLUXBUGS-1614): Stop passing binary image to IIB
IIB has its own internal mapping of binary images for each OCP version.
If we set the version ourselves then it will override what IIB does.
This creates an issue where we try to consistently set the version for
the OCP tag as we can try to pass an invalid pullspec.

Signed-off-by: arewm <arewm@users.noreply.github.com>